+++ src/Type.h
@@ -347,8 +347,8 @@ inline Value  operator+  (i32 i, Value v) { return Value(i + i32(v)); }
 inline Value  operator-  (i32 i, Value v) { return Value(i - i32(v)); }
 inline Value  operator/  (Value  v, i32 i) { return Value(i32(v) / i); }
 inline Value& operator/= (Value &v, i32 i) { v = Value(i32(v) / i); return v; }
-inline Value  operator*  (Value  v, float f) { return Value(i32(i32(v) * f)); }
-inline Value& operator*= (Value &v, float f) { v = Value(i32(i32(v) * f)); return v; }
+inline Value  operator*  (Value  v, double f) { return Value(i32(i32(v) * f)); }
+inline Value& operator*= (Value &v, double f) { v = Value(i32(i32(v) * f)); return v; }
 inline i32    operator/  (Value v1, Value v2) { return i32(v1) / i32(v2); }
 
 // Make score from mid and end values
@@ -369,8 +369,8 @@ ARTHMAT_OPERATORS (Score)
 inline Score  operator*  (Score s1, Score s2);
 /// Division of a Score must be handled separately for each term
 inline Score  operator/  (Score  s, i32   i) { return mk_score (mg_value (s) / i, eg_value (s) / i); }
-inline Score  operator*  (Score  s, float f) { return mk_score (mg_value (s) * f, eg_value (s) * f); }
-inline Score& operator*= (Score &s, float f) { s = mk_score (mg_value (s) * f, eg_value (s) * f); return s; }
+inline Score  operator*  (Score  s, double f) { return mk_score (mg_value (s) * f, eg_value (s) * f); }
+inline Score& operator*= (Score &s, double f) { s = mk_score (mg_value (s) * f, eg_value (s) * f); return s; }
 
 ARTHMAT_OPERATORS (Depth)
 INC_DEC_OPERATORS (Depth)
@@ -497,8 +497,8 @@ inline Move mk_move<PROMOTE> (Square org, Square dst, PieceT pt) { return Move(P
 
 inline Move mk_move (Square org, Square dst) { return mk_move<NORMAL> (org, dst); }
 
-inline float value_to_cp (Value  v) { return (float)     v / i32(VALUE_EG_PAWN); }
-inline Value cp_to_value (float cp) { return (Value)i32(cp * i32(VALUE_EG_PAWN)); }
+inline double value_to_cp (Value   v) { return (double)     v / i32(VALUE_EG_PAWN); }
+inline Value  cp_to_value (double cp) { return (Value)i32(cp * i32(VALUE_EG_PAWN)); }
 
 inline Value mates_in (i32 ply) { return +VALUE_MATE - ply; }
 inline Value mated_in (i32 ply) { return -VALUE_MATE + ply; }
